+++ src/backup.rs
@@ -71,7 +71,10 @@ impl BackupWriter {
         })
     }
 
-    fn push_entry(&mut self, index_entry: IndexEntry) -> Result<()> {
+    /// Push a new entry into the backup's IndexBuilder.
+    ///
+    /// This is public only to facilitate testing.
+    pub(crate) fn push_entry(&mut self, index_entry: IndexEntry) -> Result<()> {
         // TODO: Return or accumulate index sizes.
         self.index_builder.push_entry(index_entry)?;
         Ok(())